исправил названия файлов, немного причесал код
