+++ src/cli/commands/generate.command.ts
