+++ src/cli/commands/version.command.ts
