+++ src/cli/index.ts
@@ -1,7 +1,7 @@
 export { CLIApplication } from './cli-application.js';
 export { CommandParser } from './command-parser.js';
 
-export { HelpCommand } from './commands/help-command.js';
-export { VersionCommand } from './commands/version-command.js';
-export { ImportCommand } from './commands/import-command.js';
-export { GenerateCommand } from './commands/generate-command.js';
+export { HelpCommand } from './commands/help.command.js';
+export { VersionCommand } from './commands/version.command.js';
+export { ImportCommand } from './commands/import.command.js';
+export { GenerateCommand } from './commands/generate.command.js';
