+++ src/main.rest.ts
@@ -110,7 +110,7 @@ bootstrap();
         или param.userId === payload.id, но это снова авторизация...
         а откуда клиент знает id пользователя, если еще не авторизован...
         глянуть какие на клиенте будут запросы...
-  37. Перепроверить описание ответов об ошибках 400 401 403... поменять описание, проверить использование ".notAllow"
+  37. Перепроверить описание ответов об ошибках 400 401 403... поменять описание, проверить использование ".notAllow" .notFound .noContent
 
 
   50. tsconfig добавил алиасы / vscode распознает пути, а копилятор нет
