+++ src/shared/libs/offer-generator/index.ts
@@ -1,2 +1,2 @@
 export { OfferGenerator } from './offer-generator.interface.js';
-export { TSVOfferGenerator } from './tsv-offer-generator.js';
+export { TSVOfferGenerator } from './tsv.offer-generator.js';
