+++ src/shared/modules/auth/index.ts
@@ -1,7 +1,7 @@
 export { BaseUserException, UserNotFoundException, UserPasswordIncorrectException } from './errors/index.js';
-export { TokenPayload } from './types/TokenPayload.js';
-export { AuthExceptionFilter } from './auth-exception-filter.js';
+export { TokenPayload } from './type/token-payload.type.js';
 export { AuthService } from './auth-service.interface.js';
 export { JWT_ALGORITHM, JWT_EXPIRED } from './auth.constant.js';
 export { createAuthContainer } from './auth.container.js';
-export { DefaultAuthService } from './default-auth.service.js';
+export { AuthExceptionFilter } from './auth.exception-filter.js';
+export { DefaultAuthService } from './default.auth-service.js';
