+++ src/shared/modules/auth/type/token-payload.type.ts
@@ -1,0 +1,7 @@
+export type TokenPayload = {
+  user: {
+    name: string;
+    email: string;
+    id: string;
+  }
+};
