--- src/shared/modules/auth/types/TokenPayload.ts
@@ -1,5 +1,0 @@
-export type TokenPayload = {
-  name: string;
-  email: string;
-  id: string;
-};
