+++ src/shared/modules/offer/default.offer-service.ts
