+++ src/shared/modules/offer/offer.container.ts
@@ -2,7 +2,7 @@ import { Container } from 'inversify';
 import { types } from '@typegoose/typegoose';
 import { OfferService } from './offer-service.interface.js';
 import { Component } from '../../types/index.js';
-import { DefaultOfferService } from './default-offer.service.js';
+import { DefaultOfferService } from './default.offer-service.js';
 import { OfferEntity, OfferModel } from './offer.entity.js';
 import { Controller } from 'src/shared/libs/rest/index.js';
 import { OfferController } from './offer.controller.js';
