+++ src/shared/modules/review/index.ts
@@ -2,7 +2,7 @@ export { CreateReviewDto } from './dto/create-review.dto.js';
 export { ReviewRdo } from './rdo/review.rdo.js';
 export { CreateReviewRequest } from './type/create-review-request.type.js';
 export { IndexReviewsRequest } from './type/index-reviews-request.type.js';
-export { DefaultReviewService } from './default-review.service.js';
+export { DefaultReviewService } from './default.review-service.js';
 export { ReviewService } from './review-service.interface.js';
 export { DEFAULT_REVIEW_COUNT, ReviewValidation, ReviewValidationMessage } from './review.const.js';
 export { createReviewContainer } from './review.container.js';
