+++ src/shared/modules/review/review.container.ts
@@ -2,7 +2,7 @@ import { Container } from 'inversify';
 import { types } from '@typegoose/typegoose';
 import { ReviewService } from './review-service.interface.js';
 import { Component } from '../../types/index.js';
-import { DefaultReviewService } from './default-review.service.js';
+import { DefaultReviewService } from './default.review-service.js';
 import { ReviewEntity, ReviewModel } from './review.entity.js';
 import { Controller } from '../../libs/rest/index.js';
 import { ReviewController } from './review.controller.js';
