+++ src/shared/modules/user/index.ts
@@ -5,7 +5,7 @@ export { UserRdo } from './rdo/user.rdo.js';
 export { CreateUserRequest } from './type/create-user-request.type.js';
 export { ParamUserId } from './type/param-user-id.type.js';
 export { LoginUserRequest } from './type/login-user-request.type.js';
-export { DefaultUserService } from './default-user.service.js';
+export { DefaultUserService } from './default.user-service.js';
 export { UserService } from './user-service.interface.js';
 export { DEFAULT_AVATAR_PATH, UserName, UserRoute, UserValidation, UserValidationMessage } from './user.const.js';
 export { createUserContainer } from './user.container.js';
