+++ src/shared/modules/user/user.container.ts
@@ -2,7 +2,7 @@ import { Container } from 'inversify';
 import { types } from '@typegoose/typegoose';
 import { UserService } from './user-service.interface.js';
 import { Component } from '../../types/index.js';
-import { DefaultUserService } from './default-user.service.js';
+import { DefaultUserService } from './default.user-service.js';
 import { UserEntity, UserModel } from './index.js';
 import { Controller } from '../../libs/rest/index.js';
 import { UserController } from './user.controller.js';
